move legacy code into benchmark test, to keep it as documentation for progressions
